+++ tests/model/providers/test_llama_cpp_python.py
@@ -1,0 +1,30 @@
+import pytest
+from test_helpers.utils import skip_if_no_llama_cpp_python
+
+from inspect_ai.model import (
+    ChatMessageUser,
+    GenerateConfig,
+    get_model,
+)
+
+
+@pytest.mark.asyncio
+@skip_if_no_llama_cpp_python
+async def test_llama_cpp_python_api() -> None:
+    model = get_model(
+        "llama-cpp-python/default",
+        config=GenerateConfig(
+            frequency_penalty=0.0,
+            stop_seqs=None,
+            max_tokens=50,
+            presence_penalty=0.0,
+            logit_bias=dict([(42, 10), (43, -10)]),
+            seed=None,
+            temperature=0.0,
+            top_p=1.0,
+        ),
+    )
+
+    message = ChatMessageUser(content="This is a test string. What are you?")
+    response = await model.generate(input=[message])
+    assert len(response.completion) >= 1